Fix inheritance of exclude by moving into non-.rubocop files
When Excludes are present in an inherited configuration file, Rubocop
will expand_path them.  However, the Excludes in .rubocop_base.yml were
being resolved relative to the gem itself.  Because of this they were
ultimately ignored.

The reason this happens is because Rubocop looks at the filename, and if
it starts with .rubocop, it assumes that is a project-level file, and
will expand relative to the file itself.  Rubocop's own default config
intentionally does _not_ start with .rubocop, as so the default config
is expanded relative to the project. See
https://github.com/rubocop-hq/rubocop/blob/eb3d3583/lib/rubocop/config.rb#L190-L203

The simplest fix for now is to just rename our files to not have
.rubocop at the start of the file name.  We can still keep our
.rubocop_base.yml to avoid renaming it in all of the gems.

Fixes #7
Closes #8