Drop ruby requirement from gemspec
The Ruby requirement will ultimately come from whatever version of
rubocop we use, so there's no need for us to pin it here. It turns out
that rubocop 1.56.3 require Ruby 2.7, but having it here makes it appear
like a decision of manageiq-style as opposed to the restriction coming
from rubocop itself.